refactor: move constants and types to statsig dir
