+++ src/statsig/types.ts
@@ -1,0 +1,11 @@
+export enum StatsigLayers {
+  NAME_AND_PICTURE_SCREEN = 'name_and_picture_screen',
+}
+
+export enum StatsigDynamicConfigs {
+  USERNAME_BLOCK_LIST = 'username_block_list',
+}
+
+export enum StatsigExperiments {
+  ADD_FUNDS_CRYPTO_EXCHANGE_QR_CODE = 'add_funds_crypto_exchange_qr_code',
+}
